requirements.txt
ipython==5.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
Jinja2==2.11.3
jsonschema==2.6.0
librosa==0.8.0
matplotlib==3.2.2
matplotlib-venn==0.11.6
numpy==1.19.5
pandas==1.1.5
pickleshare==0.7.5
python-apt==0.0.0
python-speech-features==0.6
python-utils==2.5.6
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.1
sklearn==0.0
sklearn-pandas==1.8.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.7
sqlparse==0.4.1
torch==1.8.1+cu101
torchsummary==1.5.1
torchtext==0.9.1
uritemplate==3.0.1
certifi==2020.6.20
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
MarkupSafe==1.1.1
Pillow==7.2.0
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
openpyxl==2.5.9